# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Knowledge Platform API",
        "description": "Read models, search and knowledge-graph APIs for the knowledge\npublishing platform.\n\nThe backend returns JSON read models only. Article bodies arrive as a\nstructured block tree rather than HTML: presentation belongs to the\nfrontend, and keeping markup out of the API is also what keeps the\nfrontend from becoming a sanitisation boundary.\n\nPublic read endpoints are anonymous. Everything under /api/studio\nrequires the AUTHOR role. Webhook endpoints are anonymous at the HTTP\nlayer and authenticated by provider signature.\n",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/studio/authors/{slug}": {
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update an author's profile",
                "operationId": "updateAuthor",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAuthorProfileRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/studio/articles/{slug}/unpublish": {
            "post": {
                "tags": [
                    "Studio: Articles"
                ],
                "summary": "Withdraw an article from publication",
                "operationId": "unpublish",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/studio/articles/{slug}/publish": {
            "post": {
                "tags": [
                    "Studio: Articles"
                ],
                "summary": "Publish an article",
                "operationId": "publish",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/studio/articles/{slug}/archive": {
            "post": {
                "tags": [
                    "Studio: Articles"
                ],
                "summary": "Archive an article, retaining it so inbound links are not orphaned",
                "operationId": "archive",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/studio/articles": {
            "get": {
                "tags": [
                    "Studio: Articles"
                ],
                "summary": "List published articles",
                "operationId": "list_1",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseArticleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/studio/articles/{slug}": {
            "get": {
                "tags": [
                    "Studio: Articles"
                ],
                "summary": "Fetch an article in its editorial form, whatever its publication state",
                "operationId": "get_2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "List authors",
                "operationId": "listAuthors",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseAuthorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/authors/{slug}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Fetch an author's public profile",
                "operationId": "getAuthor",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/articles": {
            "get": {
                "tags": [
                    "Delivery"
                ],
                "summary": "The article archive, newest first",
                "operationId": "articles",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseArticleReference"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{slug}": {
            "get": {
                "tags": [
                    "Delivery"
                ],
                "summary": "The complete article page read model",
                "description": "Includes body blocks, outline, author, taxonomy, related articles, backlinks,\nbreadcrumbs, previous/next navigation and a bounded graph neighbourhood --\neverything a normal article page needs, with no follow-up calls.\n",
                "operationId": "article",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleReadModel"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UpdateAuthorProfileRequest": {
                "type": "object",
                "properties": {
                    "displayName": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    },
                    "biography": {
                        "type": "string",
                        "maxLength": 10000,
                        "minLength": 0
                    },
                    "avatarUrl": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 0
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 320,
                        "minLength": 0
                    }
                },
                "required": [
                    "displayName"
                ]
            },
            "AuthorResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "biography": {
                        "type": "string"
                    },
                    "avatarUrl": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ArticleResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "publicationState": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    },
                    "authorId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "topics": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "wordCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "readingTimeMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "revisionNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sourcePath": {
                        "type": "string"
                    },
                    "sourceRevision": {
                        "type": "string"
                    },
                    "publishedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ArticleReference": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "authorName": {
                        "type": "string"
                    }
                }
            },
...
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "PageResponseArticleResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleResponse"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
...
            "PageResponseAuthorResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuthorResponse"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
...
            "PageResponseArticleReference": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleReference"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
            "ArticleReadModel": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "body": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/CodeBlock"
                                },
                                {
                                    "$ref": "#/components/schemas/Heading"
                                },
                                {
                                    "$ref": "#/components/schemas/ListBlock"
                                },
                                {
                                    "$ref": "#/components/schemas/Paragraph"
                                },
                                {
                                    "$ref": "#/components/schemas/Quote"
                                },
                                {
                                    "$ref": "#/components/schemas/TableBlock"
                                },
                                {
                                    "$ref": "#/components/schemas/ThematicBreak"
                                }
                            ]
                        }
                    },
                    "outline": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DocumentHeading"
                        }
                    },
                    "author": {
                        "$ref": "#/components/schemas/AuthorSummaryModel"
                    },
                    "topics": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaxonomyReference"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaxonomyReference"
                        }
                    },
                    "relatedArticles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleReference"
                        }
                    },
                    "backlinks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleReference"
                        }
                    },
                    "unresolvedReferences": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "breadcrumbs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BreadcrumbEntry"
                        }
                    },
                    "previousArticle": {
                        "$ref": "#/components/schemas/ArticleReference"
                    },
                    "nextArticle": {
                        "$ref": "#/components/schemas/ArticleReference"
                    },
                    "graphContext": {
                        "$ref": "#/components/schemas/GraphNeighbourhood"
                    },
                    "wordCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "readingTimeMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "publishedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "CodeBlock": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ContentBlock"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "language": {
                                "type": "string"
                            },
                            "code": {
                                "type": "string"
                            }
                        }
                    }
                ]
            },
...
            "Heading": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ContentBlock"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "level": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "anchorId": {
                                "type": "string"
                            },
                            "text": {
                                "type": "string"
                            },
                            "content": {
                                "type": "array",
                                "items": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ExternalLink"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ImageReference"
                                        },
                                        {
                                            "$ref": "#/components/schemas/InlineCode"
                                        },
                                        {
                                            "$ref": "#/components/schemas/InternalLink"
                                        },
                                        {
                                            "$ref": "#/components/schemas/LineBreak"
                                        },
                                        {
                                            "$ref": "#/components/schemas/StyledRun"
                                        },
                                        {
                                            "$ref": "#/components/schemas/TextRun"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                ]
            },
...
            "ListBlock": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ContentBlock"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "ordered": {
                                "type": "boolean"
                            },
                            "startNumber": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ListItem"
                                }
                            }
                        }
                    }
                ]
            },
...
            "Paragraph": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ContentBlock"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "content": {
                                "type": "array",
                                "items": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ExternalLink"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ImageReference"
                                        },
                                        {
                                            "$ref": "#/components/schemas/InlineCode"
                                        },
                                        {
                                            "$ref": "#/components/schemas/InternalLink"
                                        },
                                        {
                                            "$ref": "#/components/schemas/LineBreak"
                                        },
                                        {
                                            "$ref": "#/components/schemas/StyledRun"
                                        },
                                        {
                                            "$ref": "#/components/schemas/TextRun"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                ]
            },
...
            "TableBlock": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ContentBlock"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "header": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/TableRow"
                                }
                            },
                            "rows": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/TableRow"
                                }
                            }
                        }
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}